pyproject.toml
[tool.poetry]
name = "django-shopper"
version = "0.1.0"
description = ""
authors = ["masoudhassanzadeh <mhz1992@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.2.2"
djangorestframework = "^3.14.0"
python-decouple = "^3.8"
drf_yasg = "^1.21.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
